+++ shuup/admin/modules/products/sections.py
@@ -20,11 +20,14 @@ class ProductOrdersSection(Section):
     template = "shuup/admin/products/_product_orders.jinja"
     order = 1
 
-    @staticmethod
-    def visible_for_object(product):
-        return bool(product.pk)
+    @classmethod
+    def visible_for_object(cls, product, request=None):
+        has_product_id = bool(product.pk)
+        if not request:
+            return has_product_id
+        return bool(request.user.has_perm('shuup.view_product') and has_product_id)
 
-    @staticmethod
-    def get_context_data(product):
+    @classmethod
+    def get_context_data(cls, product, request=None):
         # TODO: restrict to first 100 orders - do pagination later
         return Order.objects.valid().filter(lines__product_id=product.id).distinct()[:100]